Fix GetDigitCount overflow on int.MinValue / long.MinValue (#340)
Math.Abs(int.MinValue) and Math.Abs(long.MinValue) throw OverflowException
because |MinValue| is not representable in the same signed type. The
existing implementation therefore threw instead of returning 10 / 19.

- Int32Extensions.GetDigitCount now widens to long before Math.Abs.
- Int64Extensions.GetDigitCount computes the absolute value as ulong
  via an unchecked negation, which correctly yields 9223372036854775808
  for long.MinValue.

Both paths preserve the (int)(uint)Math.Log10(...) cast chain so that
the zero case (Log10(0) = -Infinity) still produces 1 via saturating
float-to-unsigned conversion in .NET 7+.

Extends the existing theory data with MinValue/MaxValue and +/-1 cases.

+++ src/DSE.Open/Int32Extensions.cs
@@ -27,7 +27,8 @@ public static class Int32Extensions
     /// <returns></returns>
     public static int GetDigitCount(this int number)
     {
-        return (int)(uint)Math.Log10(Math.Abs(number)) + 1;
+        // Widen to long so that Math.Abs(int.MinValue) does not overflow.
+        return (int)(uint)Math.Log10(Math.Abs((long)number)) + 1;
     }
 
     public static ulong GetRepeatableHashCode(this int number)
